move editor logic from HomePage to TextEditor
